Move SoundRecorder, SoundBufferRecorder and co. into `capture` module
